feat(ticketing): move bonus checkbox to payment screen
Move PayWithBonusPointsCheckbox from purchase overview to purchase
confirmation screen, placed above the pay button. Selection is now
stateful on the confirmation screen so toggling bonus triggers a new
offer search and updates the displayed price in real time.
